Fix click / touch events / scrolling on Windows (#6)
On Windows, click events do not come through most of the time.
Also, scrolling does not work like it does on Linux.  On Linux, you can
click and drag and scroll.  Also, on Linux, we see touch events but
no mouse events.

To receive touch events and not mouse events, use emulateTouch = true.
To fix event handling, do not send touch move events if the finger
is not in contact with the surface.  This is emulated by using SetCapture
to capture mouse handling on mouse down.  On mouse up, ReleaseCapture
releases mouse capture.  On move events, check that we have capture,
which tells us if mouse is down.